# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: waste-events
    description: Food waste event operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /waste-events:
    get:
      operationId: listWasteEvents
      tags: [waste-events]
      summary: List waste events
      description: Returns waste events, optionally filtered by date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            description: ISO 8601 datetime for start of range
        - name: to
          in: query
          required: false
          schema:
            type: string
            description: ISO 8601 datetime for end of range
        - name: station
          in: query
          required: false
          schema:
            type: string
            enum:
              - Sauté
              - Salad
              - Prep
              - Fry
              - Flat
            description: Filter by station name
      responses:
        "200":
          description: List of waste events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteEvent"
    post:
      operationId: createWasteEvent
      tags: [waste-events]
      summary: Create a waste event
      description: Log a new food waste event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWasteEventBody"
      responses:
        "201":
          description: Created waste event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteEvent"
        "400":
          description: Invalid request body

  /waste-events/summary:
    get:
      operationId: getWasteEventsSummary
      tags: [waste-events]
      summary: Get waste analytics summary
      description: Returns aggregated waste data for analytics
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            description: ISO 8601 datetime for start of range
        - name: to
          in: query
          required: false
          schema:
            type: string
            description: ISO 8601 datetime for end of range
      responses:
        "200":
          description: Aggregated waste summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteEventsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    WasteEvent:
      type: object
      properties:
        id:
          type: integer
        station:
          type: string
        wasteReason:
          type: string
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
          format: date-time
      required:
        - id
        - station
        - wasteReason
        - notes
        - recordedAt

    CreateWasteEventBody:
      type: object
      properties:
        station:
          type: string
          enum:
            - Sauté
            - Salad
            - Prep
            - Fry
            - Flat
        wasteReason:
          type: string
          enum:
            - Burnt food
            - Dropped
            - Prep mistake
            - Expired
        notes:
          type: string
      required:
        - station
        - wasteReason

    StationCount:
      type: object
      properties:
        station:
          type: string
        count:
          type: integer
      required:
        - station
        - count

    ReasonCount:
      type: object
      properties:
        wasteReason:
          type: string
        count:
          type: integer
      required:
        - wasteReason
        - count

    DailyCount:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count

    WasteEventsSummary:
      type: object
      properties:
        totalCount:
          type: integer
        byStation:
          type: array
          items:
            $ref: "#/components/schemas/StationCount"
        byReason:
          type: array
          items:
            $ref: "#/components/schemas/ReasonCount"
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/DailyCount"
      required:
        - totalCount
        - byStation
        - byReason
        - byDay