# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns waste events, optionally filtered by date range
 * @summary List waste events
 */
export const ListWasteEventsQueryParams = zod.object({
  from: zod.coerce.string().optional(),
  to: zod.coerce.string().optional(),
  station: zod.enum(["Sauté", "Salad", "Prep", "Fry", "Flat"]).optional(),
});

export const ListWasteEventsResponseItem = zod.object({
  id: zod.number(),
  station: zod.string(),
  wasteReason: zod.string(),
  notes: zod.string().nullable(),
  recordedAt: zod.coerce.date(),
});
export const ListWasteEventsResponse = zod.array(ListWasteEventsResponseItem);

/**
 * Log a new food waste event
 * @summary Create a waste event
 */
export const CreateWasteEventBody = zod.object({
  station: zod.enum(["Sauté", "Salad", "Prep", "Fry", "Flat"]),
  wasteReason: zod.enum(["Burnt food", "Dropped", "Prep mistake", "Expired"]),
  notes: zod.string().optional(),
});

/**
 * Returns aggregated waste data for analytics
 * @summary Get waste analytics summary
 */
export const GetWasteEventsSummaryQueryParams = zod.object({
  from: zod.coerce.string().optional(),
  to: zod.coerce.string().optional(),
});

export const GetWasteEventsSummaryResponse = zod.object({
  totalCount: zod.number(),
  byStation: zod.array(
    zod.object({
      station: zod.string(),
      count: zod.number(),
    }),
  ),
  byReason: zod.array(
    zod.object({
      wasteReason: zod.string(),
      count: zod.number(),
    }),
  ),
  byDay: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
    }),
  ),
});